unsigned -> size_t

+++ include/output.h
@@ -5,6 +5,6 @@
 
 void print_text          (struct Text_t text);
 void print_str           (char str_name[], const char *string);
-void print_pointer_array (char **ptr_array, unsigned n_lines);
+void print_pointer_array (char **ptr_array, size_t n_lines);
 
 #endif //OUTPUT_H
